+++ packages/tasks_panel/src/app-events.ts
@@ -1,7 +1,7 @@
 import {
   ConfiguredTask,
   TaskEditorContributionAPI,
-} from "@vscode-tasks-explorer/task_contrib_types";
+} from "@sap_oss/task_contrib_types";
 
 /**
  * Interface AppEvents