+++ packages/tasks_panel/src/utils/task-serializer.ts
@@ -1,4 +1,4 @@
-import { ConfiguredTask } from "@vscode-tasks-explorer/task_contrib_types";
+import { ConfiguredTask } from "@sap_oss/task_contrib_types";
 
 export function serializeTask(task: ConfiguredTask): string {
   return JSON.stringify(task);