+++ packages/tasks_panel/src/utils/ws-folder.ts
@@ -1,5 +1,5 @@
 import { forEach } from "lodash";
-import { ConfiguredTask } from "@vscode-tasks-explorer/task_contrib_types";
+import { ConfiguredTask } from "@sap_oss/task_contrib_types";
 
 // this function removes "index" field that Theia adds to the task definition
 export function cleanTasks(tasks: ConfiguredTask[]): void {