+++ packages/tasks_panel/test/utils/mockTasksProvider.ts
@@ -1,4 +1,4 @@
-import { ConfiguredTask } from "@vscode-tasks-explorer/task_contrib_types";
+import { ConfiguredTask } from "@sap_oss/task_contrib_types";
 import {
   ITasksEventHandler,
   ITasksProvider,