+++ packages/task_contrib_types/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.4](https://github.com/SAP/task-explorer/compare/v3.0.3...v3.0.4) (2021-04-12)
+
+**Note:** Version bump only for package @sap_oss/task_contrib_types
+
+## [3.0.3](https://github.com/SAP/task-explorer/compare/v3.0.2...v3.0.3) (2021-04-07)
+
+**Note:** Version bump only for package @sap_oss/task_contrib_types
+
 ## [3.0.2](https://github.com/SAP/task-explorer/compare/v3.0.1...v3.0.2) (2021-04-07)
 
 ### Bug Fixes
@@ -11,7 +19,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [3.0.1](https://github.com/SAP/task-explorer/compare/v3.0.0...v3.0.1) (2021-04-06)
 
-**Note:** Version bump only for package @vscode-tasks-explorer/task_contrib_types
+**Note:** Version bump only for package @sap_oss/task_contrib_types
 
 # 3.0.0 (2021-04-04)